remove console.logs and rename file to plausibleUtils instead of analytics
